Handle volume prices without roles correctly.
Volume prices by default are not associated with any role. Users however
in old systems are associated with the "User" role by default.

This means that when a user is logged in, they will never get prices
that don't require a role.

Instead, if a user is logged in, they should see prices that are non
role specific, and prices that are.